* VAction (StyleSwitch.caseTransitionUsage): Added. * VActionMembers (addNode): If the name of the feature f is null, use "noname" instead. (caseDecisionNode): Added. * VBehavior (LineFoldStringBuilder, convertToDescription, caseTransitionUsage): Moved from VStateMembers. * VStateMembers (LineFoldStringBuilder, convertToDescription, caseTransitionUsage): Moved to VBehavior.
