+++ org.omg.sysml.plantuml/src/org/omg/sysml/plantuml/VBehavior.java
@@ -32,6 +32,7 @@ import org.omg.sysml.lang.sysml.ActionUsage;
 import org.omg.sysml.lang.sysml.Element;
 import org.omg.sysml.lang.sysml.Expression;
 import org.omg.sysml.lang.sysml.Feature;
+import org.omg.sysml.lang.sysml.FeatureMembership;
 import org.omg.sysml.lang.sysml.FeatureReferenceExpression;
 import org.omg.sysml.lang.sysml.FeatureValue;
 import org.omg.sysml.lang.sysml.Membership;
@@ -39,7 +40,10 @@ import org.omg.sysml.lang.sysml.ReferenceUsage;
 import org.omg.sysml.lang.sysml.SendActionUsage;
 import org.omg.sysml.lang.sysml.StateSubactionKind;
 import org.omg.sysml.lang.sysml.StateSubactionMembership;
+import org.omg.sysml.lang.sysml.Step;
 import org.omg.sysml.lang.sysml.Succession;
+import org.omg.sysml.lang.sysml.TransitionFeatureMembership;
+import org.omg.sysml.lang.sysml.TransitionUsage;
 
 public abstract class VBehavior extends VDefault {
     protected boolean isDoneAction(Feature f) {
@@ -159,6 +163,117 @@ public abstract class VBehavior extends VDefault {
         return null;
     }
 
+
+    private static class LineFoldStringBuilder {
+    	private StringBuilder sb = new StringBuilder();
+        private static int LINEFOLD_LEN = 20;
+        private int curLen;
+        public void fold() {
+            if (curLen <= LINEFOLD_LEN) return;
+            sb.append("\\n");
+            curLen = 0;
+        }
+
+        public void append(String s) {
+            curLen += s.length();
+            sb.append(s);
+        }
+
+        public void append(char ch) {
+            curLen++;
+            sb.append(ch);
+        }
+        
+        @Override
+        public String toString() {
+        	return sb.toString();
+        }
+    }
+
+    private String triggerToText(Step s) {
+        Visitor v = new Visitor(this) {
+            public String caseFeature(Feature f) {
+                addFeatureTypeText("", f);
+                return "";
+            }
+        };
+        for (Feature f: s.getParameter()) {
+            v.visit(f);
+        }
+        String str = v.getString();
+        if (str.isEmpty()) {
+            return getText(s);
+        } else {
+            return str;
+        }
+    }
+
+    private String convertToDescription(TransitionUsage tu) {
+        String triggerString = null;
+        String guardString = null;
+        String effectString = null;
+
+        for (FeatureMembership fm: tu.getOwnedFeatureMembership()) {
+            if (!(fm instanceof TransitionFeatureMembership)) continue;
+            TransitionFeatureMembership tfm = (TransitionFeatureMembership) fm;
+            Step s = tfm.getTransitionFeature();
+            if (s == null) continue;
+            switch (tfm.getKind()) {
+            case TRIGGER:
+                triggerString = triggerToText(s);
+                break;
+            case GUARD:
+                guardString = getText(s);
+                break;
+            case EFFECT:
+                effectString = getText(s);
+                break;
+            }
+        }
+
+        LineFoldStringBuilder ls = new LineFoldStringBuilder();
+        if (triggerString != null) {
+            triggerString = triggerString.trim();
+            if (!(triggerString.isEmpty())) {
+                ls.append(triggerString);
+                ls.append(' ');
+            }
+        }
+        if (guardString != null) {
+            guardString = guardString.trim();
+            if (!guardString.isEmpty()) {
+                ls.fold();
+                ls.append('[');
+                ls.append(guardString);
+                ls.append(']');
+            }
+        }
+        if (effectString != null) {
+            effectString = effectString.trim();
+            if (!effectString.isEmpty()) {
+                ls.fold();
+                ls.append('/');
+                ls.append(effectString);
+            }
+        }
+
+        return ls.toString();
+    }
+
+    @Override
+    public String caseTransitionUsage(TransitionUsage tu) {
+        String description = convertToDescription(tu);
+        Feature src = tu.getSource();
+        Feature tgt = tu.getTarget();
+    	if (isDoneAction(tgt)) tgt = null;
+        if ((src == null) && (tgt == null)) return "";
+        if ((src == null) || (tgt == null)) {
+            addEntryExitTransitions(new PRelation(src, tgt, tu, description));
+        } else {
+            addPRelation(src, tgt, tu, description);
+        }
+        return "";
+    }
 
     VBehavior(Visitor v) {
         super(v);
